% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_time.R
\name{discretize_time}
\alias{discretize_time}
\title{Takes a dataset with continuous time-to-event variables and target times at which to estimate the threshold-response cumulative incidence
and returns a new dataset with the time-to-event variables discretized into integer values and the new target integer times that correspond with the continuous target times.}
\usage{
discretize_time(
  data,
  Ttilde,
  biased_sampling_indicator = NULL,
  target_times,
  nbins_t = 25
)
}
\arguments{
\item{data}{A data.frame containing the observed data.
Its columns should include the covariates for which to adjust, the treatment variable,
the time-to-event-or-censoring variable `Ttilde := min(T, C)`, the indicator of whether the survival event is observed `Delta := 1(C <= T)`.}

\item{Ttilde}{A string giving the column name of the time-to-event-or-censoring variable `Ttilde := min(T, C)`.}

\item{biased_sampling_indicator}{A string giving the column name of the indicator variable of whether or not the observation's treatment is observed.
The indicator should take the value `1` if the treatment is observed and `0` otherwise.
This parameter combined with the weight parameter \code{weights_var} allows this method to be used with biased sampling designs.}

\item{target_times}{A numeric vector of time-points at which to estimate the cumulative incidence.}

\item{nbins_t}{The number of bins to use for discretizing time. The recommended value is 20-30.}
}
\description{
Takes a dataset with continuous time-to-event variables and target times at which to estimate the threshold-response cumulative incidence
and returns a new dataset with the time-to-event variables discretized into integer values and the new target integer times that correspond with the continuous target times.
}
