% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmleThresh.R
\name{pointTreatmentThresh}
\alias{pointTreatmentThresh}
\title{Main function for the threshold-response Targeted Maximum Likelihood Estimator (TMLE).}
\usage{
pointTreatmentThresh(
  data_full,
  node_list,
  thresholds = NULL,
  biased_sampling_strata = NULL,
  biased_sampling_indicator = NULL,
  lrnr_A = Lrnr_glmnet$new(),
  lrnr_Y = Lrnr_glmnet$new(),
  lrnr_Delta = Lrnr_glmnet$new()
)
}
\arguments{
\item{data_full}{A data.frame or data.table containing the data.
It should include columns for the baseline varibles W, treatment A, outcome Y, Missingness Delta (if applicable), and weights and variables for biased sampling (if applicable)}

\item{node_list}{A named list that maps nodes to variable name(s) in data Valid nodes include "W" (for baseline variables), "A" (for treatment), "Y" (for outcome variable), "Delta" (For outcome missingness), "weights"  (to adjust for biased sampling)
Example: node_list <- list(W = c("Age", "Height", "Weight"), A = "Calcium_level", Y = "indicator_event", Delta = "Missing_indicator", weights = "sampling_weights")}

\item{thresholds}{A vector of threshold values for which to estimate the upper threshold-response function EWE[Y|A >= v, W]. Including too many will slow computer time. 10-20 should be more thn enough threshold values.}

\item{biased_sampling_strata}{Variable in data that encodes grouping under which biased sampling is performed. For example, if cumulative case control sampling occurs then this variale should be the outcome variable corresponding with the node "Y".
For two-stage sampling, this should be the variable which encodes which sampling group each person belongs to. This assumes that the biased sampling is done by drawing (with or without replacement) from each of these groups.}

\item{biased_sampling_indicator}{Indicator variable in data which takes the value 1 if the individual was selected in the biased sample. Those with indicator = 0 were not selected for the biased sample and thus will not be included in the estimation procedure.}

\item{Lrnr_A}{A \code{sl3} package binomial Learner for estimating the P(A>=v|W) (make sure it supports weights if needed)}

\item{Lrnr_Y}{A \code{sl3} package Learner for estimating the conditional mean of Y (make sure it supports weights if needed)}

\item{Lrnr_Delta}{A \code{sl3} package binomial Learner for estimating the probability distribution of the missingness mechanism  (make sure it supports weights if needed)
Note that default of all learners is lasso using glmnet.}
}
\value{
In the case where only thresholds_above or thresholds_below is supplied, it is a list containing an element "results" with estimates, CI, and simultaneous CI
In the case where both thresholds_above and thresholds_below are supplied, it is a list with two elements "above_estimates" and "lower_estimates" which contains the same information as above for the upper and lower thresholds parameters.
}
\description{
Main function for the threshold-response Targeted Maximum Likelihood Estimator (TMLE).
}
