% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_functions.R
\name{sequential_targeting}
\alias{sequential_targeting}
\title{Performs the sequential regression targeting step for the threshold}
\usage{
sequential_targeting(
  data,
  data_orig,
  data_full,
  fits,
  likelihoods,
  target_times,
  node_list,
  n_full_sample = NULL,
  biased_sampling_group = NULL,
  biased_sampling_indicator = NULL,
  fast_analysis = F
)
}
\arguments{
\item{data}{The data in long format}

\item{data_orig}{The data in short format}

\item{data_full}{The data in short format including everyone with missing A values}

\item{fits}{The list of learner fits the likelihood}

\item{likelihoods}{The list of likelihoods}

\item{target_times}{The times at which to estimte the cumulative incidence}

\item{node_list}{A list named list of nodes/variable dictionary}

\item{biased_sampling_group}{The variable name for the biased sampling group}

\item{biased_sampling_indicator}{The variable name for the Indicator whether you are selected by the biased sample}

\item{fast_analysis}{Whether to use a shortcut for a faster analysis}
}
\description{
Performs the sequential regression targeting step for the threshold
}
