% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_functions.R
\name{target_J}
\alias{target_J}
\title{Targets the J component of the likelihood for the parameter of interest}
\usage{
target_J(likelihoods, fits, data, target_times, node_list)
}
\arguments{
\item{likelihoods}{The list of likelihood}

\item{fits}{The list of learner fits the likelihood}

\item{data}{The data in long format}

\item{target_times}{The times at which to estimte the cumulative incidence}

\item{node_list}{A list named list of nodes/variable dictionary}
}
\description{
Targets the J component of the likelihood for the parameter of interest
}
