% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_functions.R
\name{target_N}
\alias{target_N}
\title{Performs one iteration of targeting for the N component of the likelihood for the parameter of interest.}
\usage{
target_N(
  data,
  likelihoods,
  fits,
  node_list,
  target_times,
  n_full_sample = NULL,
  max_eps = NULL,
  force_converge = FALSE
)
}
\arguments{
\item{data}{The data in long format}

\item{likelihoods}{The list of likelihoods}

\item{fits}{The list of learner fits the likelihood}

\item{target_times}{The times at which to estimte the cumulative incidence}

\item{n_full_sample}{The total sample size (including those with missing A values)}

\item{max_eps}{Max epsilon value to consider when performing MLE}

\item{force_converge}{Forces convergence and computes EIF components. This is necessary if for whatever reason there is no convergence before the max number of iteratons.}
}
\description{
Performs one iteration of targeting for the N component of the likelihood for the parameter of interest.
}
