% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_functions.R
\name{update_J}
\alias{update_J}
\title{Updates the J component of the likelihood for the parameter of interest.
Requires that the likelihood has already been targeted by a call to \code{target_J}}
\usage{
update_J(likelihoods, fits, data, target_times)
}
\arguments{
\item{likelihoods}{The list of likelihoods}

\item{fits}{The list of learner fits the likelihood}

\item{data}{The data in long format}

\item{target_times}{The times at which to estimte the cumulative incidence}
}
\description{
Updates the J component of the likelihood for the parameter of interest.
Requires that the likelihood has already been targeted by a call to \code{target_J}
}
