% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting_functions.R
\name{update_N}
\alias{update_N}
\title{Performs one iteration of targeting for the N component of the likelihood for the parameter of interest.
Requires that the likelihood has already been targeted by a call to \code{target_N}}
\usage{
update_N(data, likelihoods, fits, node_list, target_times, step)
}
\arguments{
\item{data}{The data in long format}

\item{likelihoods}{The list of likelihoods}

\item{fits}{The list of learner fits the likelihood}

\item{node_list}{A list named list of nodes/variable dictionary}

\item{target_times}{The times at which to estimte the cumulative incidence}

\item{step}{Step number}
}
\description{
Performs one iteration of targeting for the N component of the likelihood for the parameter of interest.
Requires that the likelihood has already been targeted by a call to \code{target_N}
}
